requirements.txt
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
daphne==2.3.0
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
oauthlib==3.0.1
psycopg2==2.8.2
PyHamcrest==1.9.0
pypiwin32==223
python3-openid==3.1.0
pytz==2019.1
pywin32==224
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
urllib3==1.24.2
zope.interface==4.6.0
